PR-08: extract supervisor/ and top-level display_override.rs
Largest single-PR extraction so far (~920 LOC). Per defect D01,
display_override.rs lands at crate root (not under supervisor/)
to break the cycle that backends would create in PR-11.

- display_override.rs (top-level, new): the display-override block
  including both test statics (X11 and Wayland) and the cfg-test/
  cfg-not(test) variants of the test setters.
- supervisor/capabilities.rs (new): detect_desktop_capabilities,
  resolve_runtime_target_for_snapshot.
- supervisor/mod.rs (new): BackendContext, BackendHandle, runtime_target
  helpers, the five run_*_backend_task adapters, start_backend,
  SupervisorState, transition_runtime_target variants,
  run_lifecycle_supervisor variants, wait/poll helpers,
  WAYLAND_CAPABILITY_RECHECK_INTERVAL.

main.rs: 4911 -> 4026 LOC. Visibility widenings per D06 + D23:
run_gnome/kde/wayland/x11, query_*_focus / query_*_active_window,
BackendExit, map_run_outcome_to_backend_exit all widened to pub(crate);
plus the unplanned session_bus_name_has_owner widening. BackendExit
and the mapping helper stay in main.rs until PR-12 moves them to
backends/mod.rs.

Verification: cargo build clean; cargo test --bin kanata-switcher
261 passed / 0 failed.
